Wire up services to view model

+++ Movies.Clients/Models/PagedResponse.cs
@@ -8,15 +8,15 @@ namespace Movies.Clients.Models
 	public class PagedResponse<T>
 	{
 		[JsonProperty("page")]
-		public long Page { get; set; }
+		public int Page { get; set; }
 
 		[JsonProperty("results")]
 		public IReadOnlyList<T> Results { get; set; }
 
 		[JsonProperty("total_results")]
-		public long TotalResults { get; set; }
+		public int TotalResults { get; set; }
 
 		[JsonProperty("total_pages")]
-		public long TotalPages { get; set; }
+		public int TotalPages { get; set; }
     }
 }
